Remove the zephyr header file from the Bluetooth global header directory
VELAPLATFO-82081

`bt_le_scan.h` is a globally exposed header file. Then, Zephyr's `#include <zephyr/bluetooth/bluetooth.h>` is declared as a private inclusion of Zephyr in CMake. However, the problem is that when third-party apps use my global `bt_le_scan.h`, the CMake system doesn't know where `zephyr/bluetooth/bluetooth.h` is and throws an error. One solution is for the third-party app to also declare a private inclusion of Zephyr in CMake, but this doesn't conform to design principles. The app only needs to be concerned with my framework layer. If the app also needs to include Zephyr's header files, then the framework layer is not properly configured. Therefore, `zephyr/bluetooth/bluetooth.h` must not be explicitly included in `bt_le_scan`.

+++ service/stacks/zephyr/sal_le_scan_interface.c
@@ -126,7 +126,7 @@ static void STACK_CALL(stop_scan)(void* args)
 bt_status_t bt_sal_le_set_scan_parameters(bt_controller_id_t id, ble_scan_params_t* params)
 {
     memset(&scan_param, 0, sizeof(scan_param));
-    scan_param.type = params->scan_type;
+    scan_param.type = (uint8_t)params->scan_type;
     scan_param.interval = params->scan_interval;
     scan_param.window = params->scan_window;
